Make encryption methods algorithm independent
