+++ powerauth-java-server/src/main/java/com/wultra/security/powerauth/app/server/service/behavior/tasks/UpgradeServiceBehavior.java
@@ -25,6 +25,7 @@ import com.wultra.security.powerauth.app.server.database.model.entity.Applicatio
 import com.wultra.security.powerauth.app.server.database.repository.ActivationRepository;
 import com.wultra.security.powerauth.app.server.database.repository.ApplicationVersionRepository;
 import com.wultra.security.powerauth.app.server.service.crypto.CryptographyServiceFactory;
+import com.wultra.security.powerauth.app.server.service.crypto.v3.EncryptionServiceEcies;
 import com.wultra.security.powerauth.app.server.service.exceptions.GenericServiceException;
 import com.wultra.security.powerauth.app.server.service.i18n.LocalizationProvider;
 import com.wultra.security.powerauth.app.server.service.model.ServiceError;
@@ -43,7 +44,6 @@ import com.wultra.security.powerauth.crypto.lib.enums.ProtocolVersion;
 import com.wultra.security.powerauth.crypto.lib.generator.HashBasedCounter;
 import com.wultra.security.powerauth.crypto.lib.model.exception.CryptoProviderException;
 import com.wultra.security.powerauth.crypto.lib.model.exception.GenericCryptoException;
-import com.wultra.security.powerauth.crypto.lib.v4.model.context.SharedSecretAlgorithm;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -66,6 +66,7 @@ public class UpgradeServiceBehavior {
     private final ActivationContextValidator activationValidator;
     private final ApplicationVersionRepository applicationVersionRepository;
     private final ActivationRepository activationRepository;
+    private final EncryptionServiceEcies encryptionService;
 
     // Helper classes
     private final ObjectMapper objectMapper;
@@ -113,7 +114,7 @@ public class UpgradeServiceBehavior {
             // Try to decrypt request data, the data must not be empty. Currently only '{}' is sent in request data. Ignore result of decryption.
             final EncryptionContext context = new EncryptionContext(protocolVersion, applicationKey, activationId, EncryptorId.UPGRADE);
             // TODO - v4 support
-            final DecryptionResult decryptionResult = cryptographyServiceFactory.getService(SharedSecretAlgorithm.EC_P256).decryptRequest(encryptedRequest, context);
+            final DecryptionResult decryptionResult = encryptionService.decryptRequest(encryptedRequest, context);
 
             // Request is valid, generate hash based counter if it does not exist yet
             final String ctrDataBase64;
