+++ powerauth-java-server/src/main/java/com/wultra/security/powerauth/app/server/service/behavior/tasks/v3/EncryptionBehaviorEcies.java
@@ -20,6 +20,7 @@ package com.wultra.security.powerauth.app.server.service.behavior.tasks.v3;
 
 import com.wultra.security.powerauth.app.server.database.model.entity.ActivationRecordEntity;
 import com.wultra.security.powerauth.app.server.service.crypto.CryptographyServiceFactory;
+import com.wultra.security.powerauth.app.server.service.crypto.v3.EncryptionServiceEcies;
 import com.wultra.security.powerauth.app.server.service.exceptions.GenericServiceException;
 import com.wultra.security.powerauth.app.server.service.i18n.LocalizationProvider;
 import com.wultra.security.powerauth.app.server.service.model.ServiceError;
@@ -33,7 +34,6 @@ import com.wultra.security.powerauth.crypto.lib.encryptor.model.EncryptorId;
 import com.wultra.security.powerauth.crypto.lib.encryptor.model.EncryptorSecrets;
 import com.wultra.security.powerauth.crypto.lib.encryptor.model.v3.EciesEncryptedRequest;
 import com.wultra.security.powerauth.crypto.lib.encryptor.model.v3.ServerEciesSecrets;
-import com.wultra.security.powerauth.crypto.lib.v4.model.context.SharedSecretAlgorithm;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -63,6 +63,7 @@ public class EncryptionBehaviorEcies {
     private final ActivationQueryService activationQueryService;
     private final ActivationContextValidator activationValidator;
     private final CryptographyServiceFactory cryptographyServiceFactory;
+    private final EncryptionServiceEcies encryptionService;
 
     /**
      * Obtain ECIES decryptor parameters to allow decryption of ECIES-encrypted messages on intermediate server.
@@ -113,7 +114,7 @@ public class EncryptionBehaviorEcies {
                 request.getTimestamp()
         );
         final EncryptionContext context = new EncryptionContext(request.getProtocolVersion(), request.getApplicationKey(), null, EncryptorId.APPLICATION_SCOPE_GENERIC);
-        final EncryptorSecrets encryptorSecrets = cryptographyServiceFactory.getService(SharedSecretAlgorithm.EC_P256).deriveSecrets(encryptedRequest, context);
+        final EncryptorSecrets encryptorSecrets = encryptionService.deriveSecrets(encryptedRequest, context);
         if (encryptorSecrets instanceof ServerEciesSecrets encryptorSecretsV3) {
             // ECIES V3.0, V3.1, V3.2
             final GetEciesDecryptorResponse response = new GetEciesDecryptorResponse();
@@ -152,7 +153,7 @@ public class EncryptionBehaviorEcies {
         activationValidator.validateActiveStatus(activation.getActivationStatus(), activation.getActivationId(), localizationProvider);
 
         final EncryptionContext context = new EncryptionContext(request.getProtocolVersion(), request.getApplicationKey(), activationId, EncryptorId.ACTIVATION_SCOPE_GENERIC);
-        final EncryptorSecrets encryptorSecrets = cryptographyServiceFactory.getService(SharedSecretAlgorithm.EC_P256).deriveSecrets(
+        final EncryptorSecrets encryptorSecrets = encryptionService.deriveSecrets(
                 new EciesEncryptedRequest(
                         temporaryKeyId,
                         ephemeralPublicKey,
