+++ powerauth-java-server/src/main/java/com/wultra/security/powerauth/app/server/service/behavior/tasks/v4/EncryptionBehaviorAead.java
@@ -19,7 +19,7 @@
 package com.wultra.security.powerauth.app.server.service.behavior.tasks.v4;
 
 import com.wultra.security.powerauth.app.server.database.model.entity.ActivationRecordEntity;
-import com.wultra.security.powerauth.app.server.service.crypto.CryptographyServiceFactory;
+import com.wultra.security.powerauth.app.server.service.crypto.v4.EncryptionServiceAead;
 import com.wultra.security.powerauth.app.server.service.exceptions.GenericServiceException;
 import com.wultra.security.powerauth.app.server.service.i18n.LocalizationProvider;
 import com.wultra.security.powerauth.app.server.service.model.ServiceError;
@@ -33,7 +33,6 @@ import com.wultra.security.powerauth.crypto.lib.encryptor.model.EncryptorId;
 import com.wultra.security.powerauth.crypto.lib.encryptor.model.EncryptorSecrets;
 import com.wultra.security.powerauth.crypto.lib.v4.encryptor.model.context.AeadSecrets;
 import com.wultra.security.powerauth.crypto.lib.v4.encryptor.model.request.AeadEncryptedRequest;
-import com.wultra.security.powerauth.crypto.lib.v4.model.context.SharedSecretAlgorithm;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -59,7 +58,7 @@ public class EncryptionBehaviorAead {
     private final LocalizationProvider localizationProvider;
     private final ActivationQueryService activationQueryService;
     private final ActivationContextValidator activationValidator;
-    private final CryptographyServiceFactory cryptographyServiceFactory;
+    private final EncryptionServiceAead encryptionService;
 
     /**
      * Extract AEAD encryptor parameters to allow decryption of AEAD-encrypted messages on intermediate server.
@@ -101,7 +100,7 @@ public class EncryptionBehaviorAead {
         );
         final EncryptionContext context = new EncryptionContext(request.getProtocolVersion(), request.getApplicationKey(), null, EncryptorId.APPLICATION_SCOPE_GENERIC);
 
-        final EncryptorSecrets encryptorSecrets = cryptographyServiceFactory.getService(SharedSecretAlgorithm.EC_P384).deriveSecrets(encryptedRequest, context);
+        final EncryptorSecrets encryptorSecrets = encryptionService.deriveSecrets(encryptedRequest, context);
         if (encryptorSecrets instanceof AeadSecrets aeadSecrets) {
             // AEAD in V4
             final ExtractEncryptorResponse response = new ExtractEncryptorResponse();
@@ -132,7 +131,7 @@ public class EncryptionBehaviorAead {
         activationValidator.validateActiveStatus(activation.getActivationStatus(), activation.getActivationId(), localizationProvider);
 
         final EncryptionContext context = new EncryptionContext(request.getProtocolVersion(), request.getApplicationKey(), activationId, EncryptorId.ACTIVATION_SCOPE_GENERIC);
-        final EncryptorSecrets encryptorSecrets = cryptographyServiceFactory.getService(SharedSecretAlgorithm.EC_P384).deriveSecrets(
+        final EncryptorSecrets encryptorSecrets = encryptionService.deriveSecrets(
                 new AeadEncryptedRequest(
                         temporaryKeyId,
                         null,
